+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - new parameter title for ER and flow diagrams.
 - new command `Add-FlowchartClass`.
 - new parameter class for `Add-FlowchartNode`.
+- new command `Add-FlowchartClick`.
 
 ### Changed